agents/shade/curation-agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.37.2
accelerate==0.26.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy
python-multipart
mcp
duckduckgo-search